handle collision

+++ src/collisionManager.hpp
@@ -24,26 +24,7 @@ private:
 public:
 
     static void handleCollision(Window* window, std::vector<std::unique_ptr<Object>>& objects) {
-        // for (size_t i = 0; i < objects.size(); i++) {
-        //     Object* a = objects[i].get();
-
-        //     if (!a->getCanCollide())
-        //         continue;
-
-        //     for (size_t j = i + 1; j < objects.size(); j++) {
-        //         Object* b = objects[j].get();
-
-        //         if (!b->getCanCollide())
-        //             continue;
-
-        //         if (checkAABB(a, b)) {
-        //             a->onCollision(window, b);
-        //             b->onCollision(window, a);
-        //         }
-        //     }
-        // }
-
-        Bound bound = {0, 0, window->Width(), (float) window->Height()};
+        Bound bound = {0, 0, (float) window->Width(), (float) window->Height()};
         Quadtree quadtree(bound);
 
         for (auto& object : objects) {
